+++ gcdt/tenkai_core.py
@@ -7,11 +7,16 @@ import time
 import tarfile
 import boto3
 import subprocess
+
 from boto3.s3.transfer import S3Transfer
 from clint.textui import colored
 
+from gcdt import monitoring
 
-def deploy(applicationName, deploymentGroupName, deploymentConfigName, bucket, pre_bundle_scripts=None):
+
+def deploy(applicationName, deploymentGroupName, deploymentConfigName, bucket,
+           slack_token=None, slack_channel='systemmessages',
+           pre_bundle_scripts=None):
     """Upload bundle and deploy to deployment group.
     This includes the bundle-action.
 
@@ -19,6 +24,8 @@ def deploy(applicationName, deploymentGroupName, deploymentConfigName, bucket, p
     :param deploymentGroupName:
     :param deploymentConfigName:
     :param bucket:
+    :param slack_token:
+    :param slack_channel:
     :return: deploymentId from create_deployment
     """
     if pre_bundle_scripts:
@@ -29,7 +36,8 @@ def deploy(applicationName, deploymentGroupName, deploymentConfigName, bucket, p
     bundlefile = bundle_revision()
     etag, version = _upload_revision_to_s3(bucket, applicationName, bundlefile)
 
-    client = boto3.client('codedeploy')
+    session = boto3.session.Session()
+    client = session.client('codedeploy')
     response = client.create_deployment(
         applicationName=applicationName,
         deploymentGroupName=deploymentGroupName,
@@ -47,6 +55,17 @@ def deploy(applicationName, deploymentGroupName, deploymentConfigName, bucket, p
         description='deploy with tenkai',
         ignoreApplicationStopFailures=True
     )
+
+    print("Deployment: {} -> URL: https://{}.console.aws.amazon.com/codedeploy/home?region={}#/deployments/{}".format(
+        response['deploymentId'],
+        session.region_name,
+        session.region_name,
+        response['deploymentId'],
+    ))
+
+    message = 'tenkai bot: deployed deployment group %s ' % deploymentGroupName
+    monitoring.slack_notification(slack_channel, message, slack_token)
+
     return response['deploymentId']
 
 
@@ -115,12 +134,14 @@ def _upload_revision_to_s3(bucket, applicationName, file):
 
     return response['ETag'], response['VersionId']
 
+
 def _execute_pre_bundle_scripts(scripts):
     for script in scripts:
         exit_code = _execute_script(script)
         if exit_code != 0:
             return exit_code
     return 0
+
 
 def _execute_script(file_name):
     if os.path.isfile(file_name):